requirements.txt
# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.5.0

# Deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
torch-ema>=0.3.0

# Medical imaging libraries
nibabel>=4.0.0
pydicom>=2.3.0

# Machine learning and data science
scikit-learn>=1.1.0
scikit-image>=0.19.0
pandas>=1.4.0
seaborn>=0.11.0

# Utilities and monitoring
tqdm>=4.64.0
wandb>=0.13.0
datasets>=2.10.0
PyYAML>=6.0
click>=8.0.0
hydra-core>=1.3.0

# Medical datasets
medmnist>=1.0.0

# Testing framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0 

# Microsoft Excel file handling
openpyxl>=3.1.5

# kagglehub
kagglehub>=0.3.13